IPayloadBuilder.java: long instead of double
According to XUMM API it should be uint64

+++ src/main/java/com/fl/xumm4j/api/builder/IPayloadBuilder.java
@@ -24,7 +24,7 @@ public interface IPayloadBuilder {
 
     boolean DEFAULT_SUBMIT_TRANSACTION = true;
     boolean DEFAULT_MULTISIGN = false;
-    double DEFAULT_EXPIRE = 1440;
+    long DEFAULT_EXPIRE = 1440;
 
     /**
      * @param identifier
